Getting Callback was already called error
